made layout its own component
