+++ src/components/layout/layout.js
@@ -1,0 +1,47 @@
+'use strict';
+
+import React from 'react';
+
+import Footer from '../footer/footer';
+import Header from '../header/header';
+import Navigation from '../navigation/navigation';
+import UserDetails from '../user-details/user-details';
+
+const Layout = React.createClass({
+  render: function() {
+    return (
+      <section>
+        <section className="row">
+
+          <div className="col-md-*">
+            <Header></Header>
+          </div>
+
+        </section>
+
+        <section className="row">
+
+          <div className="col-md-3">
+            <UserDetails></UserDetails>
+            <Navigation></Navigation>
+          </div>
+
+          <div className="col-md-9">
+            <div>{this.props.children}</div>
+          </div>
+
+        </section>
+
+        <section className="row">
+
+          <div className="col-md-12">
+            <Footer></Footer>
+          </div>
+        </section>
+
+      </section>
+    )
+  }
+});
+
+export default Layout;
